Crate demo_features

Structs

AtlasDemoMeta ğŸ”’

Atlas information.

GameMeta ğŸ”’

Create our root asset type.

PersistedTextData ğŸ”’

TileMeta ğŸ”’

Tile info.

TilemapDemoMeta ğŸ”’

Tilemap info.

Functions

atlas_demo_plugin ğŸ”’

Plugin for running the atlas demo.

atlas_demo_startup ğŸ”’

System to startup the atlas demo.

audio_demo_plugin ğŸ”’

audio_demo_ui ğŸ”’

back_to_menu_ui ğŸ”’

Simple UI system that shows a button at the bottom of the screen to delete the current session and go back to the main menu.

create_game

demo_widget ğŸ”’

This is an example widget system.

main ğŸ”’

menu_plugin

Menu plugin

menu_startup ğŸ”’

Setup the main menu.

menu_system ğŸ”’

Our main menu system.

move_sprite ğŸ”’

path2d_demo_plugin ğŸ”’

path2d_demo_startup ğŸ”’

sprite_demo_plugin ğŸ”’

Plugin for running the sprite demo.

sprite_demo_startup ğŸ”’

System that spawns the sprite demo.

storage_demo_plugin ğŸ”’

storage_demo_ui ğŸ”’

tilemap_demo_plugin ğŸ”’

Plugin for running the tilemap demo.

tilemap_startup_system ğŸ”’

System for starting up the tilemap demo.